initial code refactoring
