+++ .gitignore
@@ -7,7 +7,7 @@ npm-debug.log
 *.dev
 .DS_Store
 .env
-
+.vscode/*
 coverage
 
 # Docker
