--- .vscode/launch.json
@@ -1,27 +1,0 @@
-{
-	// Use IntelliSense to learn about possible attributes.
-	// Hover to view descriptions of existing attributes.
-	// For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-	"version": "0.2.0",
-	"configurations": [
-		{
-			"type": "node",
-			"request": "launch",
-			"name": "Launch Program",
-			"skipFiles": [
-				"<node_internals>/**"
-			],
-			"program": "${workspaceFolder}/index.js",
-			"env": {
-				"CORS_WHITELIST": "http://localhost:3000,http://localhost:5000",
-				"GOOGLE_OAUTH_CLIENT_ID": "704282885018-9vrt6glo5h2f3bg5c13999rkle38ul9s.apps.googleusercontent.com",
-				"GOOGLE_OAUTH_CLIENT_SECRET": "Xy54o7Oj6ag7-CGdvRbtryp9",
-				"GOOGLE_OAUTH_CALLBACK_URL": "http://localhost:5000/auth/google/callback",
-				"JWT_SECRET": "3f7bdf6d659cf0307bb15b446285d4e1",
-				"MONGODB_URI": "mongodb://heroku_5b5l4cl7:v57dald0ddgjlue0vv16q8o48p@ds211875.mlab.com:11875/heroku_5b5l4cl7",
-				"GOOGLE_OAUTH_SUCCESS_REDIRECT_URL": "http://localhost:3000/",
-				"GOOGLE_OAUTH_FAILURE_REDIRECT_URL": "http://localhost:3000/error"
-			},
-		}
-	]
-}
