+++ src/database/index.js
@@ -1,12 +1,15 @@
 const mongoose = require('mongoose');
 
-mongoose.connect(process.env.MONGODB_URI, { useNewUrlParser: true });
+const connectDatabase = () => {
+    mongoose.connect(process.env.MONGODB_URI, { useNewUrlParser: true });
+    const db = mongoose.connection;
+    db.on('error', console.error.bind(console, 'connection error:'));
+    db.once('open', () => {
+        // we're connected
+        console.log('I am connected');
+    });
+};
 
-const db = mongoose.connection;
-db.on('error', console.error.bind(console, 'connection error:'));
-db.once('open', () => {
-    // we're connected
-    console.log('I am connected');
-});
-
-module.exports = db;
+module.exports = {
+    connectDatabase,
+};
