+++ src/routes/index.js
@@ -1,11 +1,6 @@
-const { login } = require('./login');
-const { logout } = require('./logout');
 const { registerRoute } = require('./register');
 
-const routes = [
-    login,
-    logout,
-];
+const routes = [];
 
 const register = (app) => {
     routes.map((route) => registerRoute(app, route));
