--- src/routes/logout.js
@@ -1,12 +1,0 @@
-const logout = {
-    verb: 'delete',
-    path: '/logout',
-    handler: (req, res) => {
-        console.log('Loggin out.');
-        res.send('Success!');
-    },
-};
-
-module.exports = {
-    logout,
-};
